// redux/types.ts
export enum ReduxActionType {
  CLOSE_MODAL = 'CLOSE_MODAL',
  FETCHED_CONTACTS = 'FETCHED_CONTACTS',
  FETCHING_CONTACTS = 'FETCHING_CONTACTS',
  FETCH_CONTACTS = 'FETCH_CONTACTS',
  GO_TO = 'GO_TO',
  LOAD_CONTACTS = 'LOAD_CONTACTS',
  OPEN_MODAL = 'OPEN_MODAL',
  RESET_CONTACTS = 'RESET_CONTACTS',
  SEARCH = 'SEARCH',
  SET_EVEN_ONLY = 'SET_EVEN_ONLY',
  SET_SEARCH = 'SET_SEARCH',
  SET_TOTAL = 'SET_TOTAL',
}